xpadneo: Skip L2CAP_CONF_UNKNWON
The current L2CAP implementation does not change any options if the other
side responds with "unknown options", but does if "unaccepted options" is
the answer. It is up to the implementation to decide on the effort spent on
config negotiations, therefore the current implementation is correct at
this point - but many devices (like Xbox One S controllers) are not usable
this way. A workaround for many users therefore is to disable_ertm, since
this is often the option which is unknown. I would prefer to try it again
with altered options instead of globally disable ERTM. In result, I suggest
the following patch. It simply adds a new case (L2CAP_CONF_UNKNOWN), which
does nothing but falling through to L2CAP_CONF_UNACCEPT.

Signed-off-by: Florian Dollinger <dollinger.florian@gmx.de>
Signed-off-by: Kai Krakow <kai@kaishome.de>