btrfs: allow a short burst of IO for probing read latency
If we do a probe to detect the current read latency of the device,
allow a short burst of IO so we don't just do single IO requests which
are probably not a realistic measurement anyways and won't have any
useful impact on the cumulative average.

Tests show that this returns device to their expected average latency
performance after some hours after a latency spike, and allows them to
become part of the round-robin again.

Signed-off-by: Kai Krakow <kai@kaishome.de>